+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/doc
 
+## 0.11.22-testing.3
+
+### Patch Changes
+
+- signer protocols are specified via config
+- Updated dependencies
+  - rocketh@0.11.22-testing.3
+
 ## 0.11.22-testing.2
 
 ### Patch Changes