+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.11.25-testing.3
+
+### Patch Changes
+
+- signer protocols are specified via config
+- Updated dependencies
+  - @rocketh/deploy@0.11.22-testing.3
+
 ## 0.11.25-testing.2
 
 ### Patch Changes