+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/verifier
 
+## 0.11.24-testing.3
+
+### Patch Changes
+
+- signer protocols are specified via config
+- Updated dependencies
+  - rocketh@0.11.22-testing.3
+
 ## 0.11.24-testing.2
 
 ### Patch Changes